+++ frontend/src/pages/ActivityFeed.tsx
@@ -4,8 +4,7 @@ import { getWeeklyStats } from '../api/stats/weeklyStats';
 import { getCalendarDay } from '../api/stats/dogCalendarDay';
 import ActivityCard from '../components/ActivityCard';
 import { RightSidebar } from '../components/stats_sidebar/RightSideBar';
-import { Activity, PaginatedActivities } from '../types/Activity';
-import { ActivityFilter } from '../types/ActivityFilter';
+import { Activity, PaginatedActivities, ActivityFilter } from '../types/Activity';
 import { DogCalendarDay } from '../types/DogCalendarDay';
 import { WeeklyStats } from '../types/WeeklyStats';
 import AddActivityButton from "../components/AddActivityButton";
