+++ frontend/src/types/Activity.ts
@@ -39,3 +39,12 @@ export interface PaginatedActivities {
   next: string | null;
   previous: string | null;
 }
+
+export interface ActivityFilter {
+  __trigger?: string
+  dog_id?: number;
+  runner_id?: number;
+  sport_id?: number;
+  start_date?: string; // ISO format
+  end_date?: string;
+};
