--- frontend/src/types/ActivityFilter.ts
@@ -1,8 +1,0 @@
-export interface ActivityFilter {
-  __trigger?: string
-  dog_id?: number;
-  runner_id?: number;
-  sport_id?: number;
-  start_date?: string; // ISO format
-  end_date?: string;
-};
